Make the new QNode default (#1807)
* Make the new QNode default

* more fixes

* more fixes

* black

* fix

* fix

* fix

* fix

* more

* fix

* fix

* more fixes

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix file mode

* changelog

* Apply suggestions from code review


* suggested changes

* increase coverage

* increase coverage

* increase coverage

* suggested changes

* more coverage

* linting

* add section to docstring

* Apply suggestions from code review


* Update

* Update

* update
