+++ pennylane/__init__.py
@@ -51,6 +51,7 @@ from pennylane.templates.subroutines import *
 from pennylane.optimize import *
 from pennylane import qaoa
 from pennylane.qnode import QNode, qnode
+import pennylane.qnode_old
 from pennylane.transforms import (
     adjoint,
     batch_params,
