+++ pennylane/transforms/specs.py
@@ -45,7 +45,7 @@ def specs(qnode, max_expansion=None):
         x = np.array([0.1, 0.2])
 
         dev = qml.device('default.qubit', wires=2)
-        @qml.qnode(dev)
+        @qml.qnode(dev, diff_method="parameter-shift", shift=np.pi / 4)
         def circuit(x, add_ry=True):
             qml.RX(x[0], wires=0)
             qml.CNOT(wires=(0,1))
@@ -55,50 +55,21 @@ def specs(qnode, max_expansion=None):
 
     >>> qml.specs(circuit)(x, add_ry=False)
     {'gate_sizes': defaultdict(int, {1: 1, 2: 1}),
-    'gate_types': defaultdict(int, {'RX': 1, 'CNOT': 1}),
-    'num_operations': 2,
-    'num_observables': 1,
-    'num_diagonalizing_gates': 0,
-    'num_used_wires': 2,
-    'depth': 2,
-    'num_device_wires': 2,
-    'device_name': 'default.qubit.autograd',
-    'diff_method': 'backprop'}
-
-    .. UsageDetails::
-
-        ``qml.specs`` can also be used with :class:`~.beta.qnode`:
-
-        .. code-block:: python3
-
-            x = np.array([0.1, 0.2])
-
-            dev = qml.device('default.qubit', wires=2)
-            @qml.beta.qnode(dev, diff_method="parameter-shift", shift=np.pi / 4)
-            def circuit(x, add_ry=True):
-                qml.RX(x[0], wires=0)
-                qml.CNOT(wires=(0,1))
-                if add_ry:
-                    qml.RY(x[1], wires=1)
-                return qml.probs(wires=(0,1))
-
-        >>> qml.specs(circuit)(x, add_ry=False)
-        {'gate_sizes': defaultdict(int, {1: 1, 2: 1}),
-         'gate_types': defaultdict(int, {'RX': 1, 'CNOT': 1}),
-         'num_operations': 2,
-         'num_observables': 1,
-         'num_diagonalizing_gates': 0,
-         'num_used_wires': 2,
-         'depth': 2,
-         'num_trainable_params': 1,
-         'num_device_wires': 2,
-         'device_name': 'default.qubit',
-         'diff_method': 'parameter-shift',
-         'expansion_strategy': 'gradient',
-         'gradient_options': {'shift': 0.7853981633974483},
-         'interface': 'autograd',
-         'gradient_fn': 'pennylane.gradients.parameter_shift.param_shift',
-         'num_gradient_executions': 2}
+     'gate_types': defaultdict(int, {'RX': 1, 'CNOT': 1}),
+     'num_operations': 2,
+     'num_observables': 1,
+     'num_diagonalizing_gates': 0,
+     'num_used_wires': 2,
+     'depth': 2,
+     'num_trainable_params': 1,
+     'num_device_wires': 2,
+     'device_name': 'default.qubit',
+     'diff_method': 'parameter-shift',
+     'expansion_strategy': 'gradient',
+     'gradient_options': {'shift': 0.7853981633974483},
+     'interface': 'autograd',
+     'gradient_fn': 'pennylane.gradients.parameter_shift.param_shift',
+     'num_gradient_executions': 2}
 
     """
 
@@ -126,14 +97,14 @@ def specs(qnode, max_expansion=None):
             dict[str, Union[defaultdict,int]]: dictionaries that contain QNode specifications
         """
         initial_max_expansion = qnode.max_expansion
-        qnode.max_expansion = max_expansion
+        qnode.max_expansion = initial_max_expansion if max_expansion is None else max_expansion
 
         try:
             qnode.construct(args, kwargs)
         finally:
             qnode.max_expansion = initial_max_expansion
 
-        if isinstance(qnode, qml.QNode):
+        if isinstance(qnode, qml.qnode_old.QNode):
             # TODO: remove when the old QNode is removed
             return qnode.specs
