+++ tests/interfaces/test_qnode_torch.py
@@ -19,7 +19,7 @@ torch = pytest.importorskip("torch", minversion="1.3")
 import numpy as np
 
 import pennylane as qml
-from pennylane import qnode, QNode
+from pennylane.qnode_old import qnode, QNode
 from pennylane.tape import JacobianTape
 from torch.autograd.functional import hessian, jacobian
 
@@ -838,7 +838,7 @@ class Test_adjoint:
 
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, diff_method="adjoint", interface="torch")
+        @qnode(dev, diff_method="adjoint", interface="torch")
         def circ(x):
             qml.RX(x[0], wires=0)
             qml.RY(x[1], wires=1)
@@ -872,7 +872,7 @@ class Test_adjoint:
 
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, diff_method="adjoint", interface="torch", adjoint_cache=True)
+        @qnode(dev, diff_method="adjoint", interface="torch", adjoint_cache=True)
         def circ(x):
             qml.RX(x[0], wires=0)
             qml.RY(x[1], wires=1)
@@ -906,7 +906,7 @@ class Test_adjoint:
 
         dev = qml.device("default.qubit", wires=1)
 
-        @qml.qnode(dev, diff_method="adjoint", interface="torch", adjoint_cache=False)
+        @qnode(dev, diff_method="adjoint", interface="torch", adjoint_cache=False)
         def circ(x):
             qml.RX(x, wires=0)
             return qml.expval(qml.PauliZ(0))
