+++ tests/tape/test_jacobian_tape.py
@@ -153,6 +153,28 @@ class TestJacobian:
         res = tape.jacobian(dev)
         assert res.shape == (4, 2)
 
+    def test_non_differentiable_state_error(self):
+        """Test error raised if attempting to differentiate a circuit
+        that returns the state."""
+        with JacobianTape() as tape:
+            qml.RX(0.543, wires=[0])
+            qml.state()
+
+        # by default all parameters are assumed to be trainable
+        with pytest.raises(ValueError, match=r"does not support circuits that return the state"):
+            tape.jacobian(None)
+
+    def test_non_differentiable_sampling_error(self):
+        """Test error raised if attempting to differentiate a circuit
+        that returns samples."""
+        with JacobianTape() as tape:
+            qml.RX(0.543, wires=[0])
+            qml.sample()
+
+        # by default all parameters are assumed to be trainable
+        with pytest.raises(qml.QuantumFunctionError, match=r"sampling can not be differentiated"):
+            tape.jacobian(None)
+
     def test_analytic_method_with_unsupported_params(self):
         """Test that an exception is raised if method="A" but a parameter
         only support finite differences"""
@@ -469,6 +491,34 @@ class TestJacobian:
         ):
             res = tape._choose_params_with_methods(diff_methods, argnum)
 
+    def test_hamiltonian_grad(self):
+        """Test that the gradient of Hamiltonians works as expected."""
+        dev = qml.device("default.qubit", wires=2)
+
+        with qml.tape.JacobianTape() as tape:
+            qml.RY(0.3, wires=0)
+            qml.RX(0.5, wires=1)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.Hamiltonian([-1.5, 2.0], [qml.PauliZ(0), qml.PauliZ(1)]))
+
+        tape.trainable_params = {2, 3}
+        res = qml.math.stack(tape.jacobian(dev)[0])
+
+        with qml.tape.JacobianTape() as tape1:
+            qml.RY(0.3, wires=0)
+            qml.RX(0.5, wires=1)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.PauliZ(0))
+
+        with qml.tape.JacobianTape() as tape2:
+            qml.RY(0.3, wires=0)
+            qml.RX(0.5, wires=1)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.PauliZ(1))
+
+        expected = qml.math.stack(qml.execute([tape1, tape2], dev, None))
+        assert np.allclose(expected, res)
+
 
 class TestJacobianIntegration:
     """Integration tests for the Jacobian method"""
@@ -744,10 +794,11 @@ class TestHessian:
             tape.hessian(None)
 
 
+@pytest.mark.parametrize("qn", [qml.qnode, qml.qnode_old.qnode])
 class TestObservableWithObjectReturnType:
     """Unit tests for differentiation of observables returning an object"""
 
-    def test_special_observable_qnode_differentiation(self):
+    def test_special_observable_qnode_differentiation(self, qn):
         """Test differentiation of a QNode on a device supporting a
         special observable that returns an object rathern than a nummber."""
 
@@ -813,12 +864,12 @@ class TestObservableWithObjectReturnType:
 
         # force diff_method='parameter-shift' because otherwise
         # PennyLane swaps out dev for default.qubit.autograd
-        @qml.qnode(dev, diff_method="parameter-shift")
+        @qn(dev, diff_method="parameter-shift")
         def qnode(x):
             qml.RY(x, wires=0)
             return qml.expval(SpecialObservable(wires=0))
 
-        @qml.qnode(dev, diff_method="parameter-shift")
+        @qn(dev, diff_method="parameter-shift")
         def reference_qnode(x):
             qml.RY(x, wires=0)
             return qml.expval(qml.PauliZ(wires=0))
