+++ tests/test_qnode.py
@@ -19,7 +19,7 @@ from scipy.sparse import coo_matrix
 
 import pennylane as qml
 from pennylane import numpy as pnp
-from pennylane.beta import qnode, QNode
+from pennylane import qnode, QNode
 from pennylane.transforms import draw
 from pennylane.tape import JacobianTape
 
@@ -229,13 +229,13 @@ class TestValidation:
         diff method."""
         dev = qml.device("default.qubit", wires=1)
 
-        mock_best = mocker.patch("pennylane.beta.QNode.get_best_method")
+        mock_best = mocker.patch("pennylane.QNode.get_best_method")
         mock_best.return_value = ("best", {}, dev)
 
-        mock_backprop = mocker.patch("pennylane.beta.QNode._validate_backprop_method")
+        mock_backprop = mocker.patch("pennylane.QNode._validate_backprop_method")
         mock_backprop.return_value = ("backprop", {}, dev)
 
-        mock_device = mocker.patch("pennylane.beta.QNode._validate_device_method")
+        mock_device = mocker.patch("pennylane.QNode._validate_device_method")
         mock_device.return_value = ("device", {}, dev)
 
         qn = QNode(dummyfunc, dev, diff_method="best")
