+++ tests/transforms/test_batch_transform.py
@@ -144,7 +144,7 @@ class TestBatchTransform:
         dev = qml.device("default.qubit", wires=2)
         cache = {}
 
-        @qml.beta.qnode(dev, max_diff=3, cache=cache)
+        @qml.qnode(dev, max_diff=3, cache=cache)
         def circuit(x):
             qml.Hadamard(wires=0)
             qml.RY(x, wires=0)
@@ -324,7 +324,7 @@ class TestBatchTransform:
         dev = qml.device("default.qubit", wires=1)
         dev = self.my_transform(dev, a, b)
 
-        @qml.beta.qnode(dev)
+        @qml.qnode(dev)
         def circuit(x):
             qml.Hadamard(wires=0)
             qml.RX(x, wires=0)
@@ -355,7 +355,7 @@ class TestBatchTransform:
         dev = qml.device("default.qubit", wires=1)
         dev = self.my_transform(a, b)(dev)
 
-        @qml.beta.qnode(dev)
+        @qml.qnode(dev)
         def circuit(x):
             qml.Hadamard(wires=0)
             qml.RX(x, wires=0)
@@ -573,8 +573,8 @@ class TestBatchTransformGradients:
         dev = qml.device("default.qubit", wires=2)
         qnode = qml.QNode(self.circuit, dev, interface="torch", diff_method=diff_method)
 
-        weights = torch.tensor([0.1, 0.2], requires_grad=True)
-        x = torch.tensor(0.543, requires_grad=True)
+        weights = torch.tensor([0.1, 0.2], requires_grad=True, dtype=torch.float64)
+        x = torch.tensor(0.543, requires_grad=True, dtype=torch.float64)
 
         res = self.my_transform(qnode, weights)(x)
         expected = self.expval(x.detach().numpy(), weights.detach().numpy())
